+++ extensions/gitpod-shared/src/features.ts
@@ -5,25 +5,16 @@
 
 // TODO get rid of loading inversify and reflect-metadata
 require('reflect-metadata');
-import { GitpodClient, GitpodServer, GitpodServiceImpl, WorkspaceInstanceUpdateListener } from '@gitpod/gitpod-protocol/lib/gitpod-service';
+import { GitpodClient, GitpodServer, GitpodServiceImpl } from '@gitpod/gitpod-protocol/lib/gitpod-service';
 import { JsonRpcProxyFactory } from '@gitpod/gitpod-protocol/lib/messaging/proxy-factory';
-import { NavigatorContext, User } from '@gitpod/gitpod-protocol/lib/protocol';
-import { Team } from '@gitpod/gitpod-protocol/lib/teams-projects-protocol';
+import { NavigatorContext } from '@gitpod/gitpod-protocol/lib/protocol';
 import { ErrorCodes } from '@gitpod/gitpod-protocol/lib/messaging/error';
 import { GitpodHostUrl } from '@gitpod/gitpod-protocol/lib/util/gitpod-host-url';
-import { ControlServiceClient } from '@gitpod/supervisor-api-grpc/lib/control_grpc_pb';
-import { InfoServiceClient } from '@gitpod/supervisor-api-grpc/lib/info_grpc_pb';
-import { WorkspaceInfoRequest, WorkspaceInfoResponse } from '@gitpod/supervisor-api-grpc/lib/info_pb';
-import { NotificationServiceClient } from '@gitpod/supervisor-api-grpc/lib/notification_grpc_pb';
+import { WorkspaceInfoRequest } from '@gitpod/supervisor-api-grpc/lib/info_pb';
 import { NotifyRequest, NotifyResponse, RespondRequest, SubscribeRequest, SubscribeResponse } from '@gitpod/supervisor-api-grpc/lib/notification_pb';
-import { PortServiceClient } from '@gitpod/supervisor-api-grpc/lib/port_grpc_pb';
-import { StatusServiceClient } from '@gitpod/supervisor-api-grpc/lib/status_grpc_pb';
 import { TasksStatusRequest, TasksStatusResponse, TaskState, TaskStatus } from '@gitpod/supervisor-api-grpc/lib/status_pb';
-import { TerminalServiceClient } from '@gitpod/supervisor-api-grpc/lib/terminal_grpc_pb';
 import { ListenTerminalRequest, ListenTerminalResponse, ListTerminalsRequest, SetTerminalSizeRequest, ShutdownTerminalRequest, Terminal as SupervisorTerminal, TerminalSize as SupervisorTerminalSize, WriteTerminalRequest } from '@gitpod/supervisor-api-grpc/lib/terminal_pb';
-import { TokenServiceClient } from '@gitpod/supervisor-api-grpc/lib/token_grpc_pb';
 import { GetTokenRequest } from '@gitpod/supervisor-api-grpc/lib/token_pb';
-import { PortVisibility } from '@gitpod/gitpod-protocol/lib/workspace-instance';
 import * as grpc from '@grpc/grpc-js';
 import * as fs from 'fs';
 import * as http from 'http';
@@ -34,201 +25,9 @@ import * as util from 'util';
 import * as vscode from 'vscode';
 import { CancellationToken, ConsoleLogger, listen as doListen } from 'vscode-ws-jsonrpc';
 import WebSocket = require('ws');
-import { BaseGitpodAnalyticsEventPropeties, GitpodAnalyticsEvent } from './analytics';
-import * as uuid from 'uuid';
-import { RemoteTrackMessage } from '@gitpod/gitpod-protocol/lib/analytics';
 import Log from './common/logger';
-import { TunnelPortRequest, TunnelVisiblity } from '@gitpod/supervisor-api-grpc/lib/port_pb';
 import { isGRPCErrorStatus } from './common/utils';
-
-export class SupervisorConnection {
-	readonly deadlines = {
-		long: 30 * 1000,
-		normal: 15 * 1000,
-		short: 5 * 1000
-	};
-	private readonly addr = process.env.SUPERVISOR_ADDR || 'localhost:22999';
-	private readonly clientOptions: Partial<grpc.ClientOptions>;
-	readonly metadata = new grpc.Metadata();
-	readonly status: StatusServiceClient;
-	readonly control: ControlServiceClient;
-	readonly notification: NotificationServiceClient;
-	readonly token: TokenServiceClient;
-	readonly info: InfoServiceClient;
-	readonly port: PortServiceClient;
-	readonly terminal: TerminalServiceClient;
-
-	constructor(
-		context: vscode.ExtensionContext
-	) {
-		this.clientOptions = {
-			'grpc.primary_user_agent': `${vscode.env.appName}/${vscode.version} ${context.extension.id}/${context.extension.packageJSON.version}`,
-		};
-		this.status = new StatusServiceClient(this.addr, grpc.credentials.createInsecure(), this.clientOptions);
-		this.control = new ControlServiceClient(this.addr, grpc.credentials.createInsecure(), this.clientOptions);
-		this.notification = new NotificationServiceClient(this.addr, grpc.credentials.createInsecure(), this.clientOptions);
-		this.token = new TokenServiceClient(this.addr, grpc.credentials.createInsecure(), this.clientOptions);
-		this.info = new InfoServiceClient(this.addr, grpc.credentials.createInsecure(), this.clientOptions);
-		this.port = new PortServiceClient(this.addr, grpc.credentials.createInsecure(), this.clientOptions);
-		this.terminal = new TerminalServiceClient(this.addr, grpc.credentials.createInsecure(), this.clientOptions);
-	}
-}
-
-type UsedGitpodFunction = ['getWorkspace', 'openPort', 'stopWorkspace', 'setWorkspaceTimeout', 'getWorkspaceTimeout', 'getLoggedInUser', 'takeSnapshot', 'waitForSnapshot', 'controlAdmission', 'sendHeartBeat', 'trackEvent', 'getTeams'];
-type Union<Tuple extends any[], Union = never> = Tuple[number] | Union;
-export type GitpodConnection = Omit<GitpodServiceImpl<GitpodClient, GitpodServer>, 'server'> & {
-	server: Pick<GitpodServer, Union<UsedGitpodFunction>>;
-};
-
-export class GitpodExtensionContext implements vscode.ExtensionContext {
-
-	readonly sessionId = uuid.v4();
-	readonly pendingActivate: Promise<void>[] = [];
-	readonly workspaceContextUrl: vscode.Uri;
-
-	constructor(
-		private readonly context: vscode.ExtensionContext,
-		readonly devMode: boolean,
-		readonly config: typeof import('./gitpod-plugin-model'),
-		readonly supervisor: SupervisorConnection,
-		readonly gitpod: GitpodConnection,
-		private readonly webSocket: Promise<ReconnectingWebSocket> | undefined,
-		readonly pendingWillCloseSocket: (() => Promise<void>)[],
-		readonly info: WorkspaceInfoResponse,
-		readonly owner: Promise<User>,
-		readonly user: Promise<User>,
-		readonly userTeams: Promise<Team[]>,
-		readonly instanceListener: Promise<WorkspaceInstanceUpdateListener>,
-		readonly workspaceOwned: Promise<boolean>,
-		readonly logger: Log,
-		readonly ipcHookCli: string | undefined
-	) {
-		this.workspaceContextUrl = vscode.Uri.parse(info.getWorkspaceContextUrl());
-	}
-
-	get active() {
-		Object.freeze(this.pendingActivate);
-		return Promise.all(this.pendingActivate.map(p => p.catch(console.error)));
-	}
-
-	get subscriptions() {
-		return this.context.subscriptions;
-	}
-	get globalState() {
-		return this.context.globalState;
-	}
-	get workspaceState() {
-		return this.context.workspaceState;
-	}
-	get secrets() {
-		return this.context.secrets;
-	}
-	get extensionUri() {
-		return this.context.extensionUri;
-	}
-	get extensionPath() {
-		return this.context.extensionPath;
-	}
-	get environmentVariableCollection() {
-		return this.context.environmentVariableCollection;
-	}
-	asAbsolutePath(relativePath: string): string {
-		return this.context.asAbsolutePath(relativePath);
-	}
-	get storageUri() {
-		return this.context.storageUri;
-	}
-	get storagePath() {
-		return this.context.storagePath;
-	}
-	get globalStorageUri() {
-		return this.context.globalStorageUri;
-	}
-	get globalStoragePath() {
-		return this.context.globalStoragePath;
-	}
-	get logUri() {
-		return this.context.logUri;
-	}
-	get logPath() {
-		return this.context.logPath;
-	}
-	get extensionMode() {
-		return this.context.extensionMode;
-	}
-	get extension() {
-		return this.context.extension;
-	}
-	get extensionRuntime() {
-		return (this.context as any).extensionRuntime;
-	}
-
-	dispose() {
-		const pendingWebSocket = this.webSocket;
-		if (!pendingWebSocket) {
-			return;
-		}
-		return (async () => {
-			try {
-				const webSocket = await pendingWebSocket;
-				await Promise.allSettled(this.pendingWillCloseSocket.map(f => f()));
-				webSocket.close();
-			} catch (e) {
-				this.logger.error('failed to dispose context:', e);
-				console.error('failed to dispose context:', e);
-			}
-		})();
-	}
-
-	async fireAnalyticsEvent({ eventName, properties }: GitpodAnalyticsEvent): Promise<void> {
-		const baseProperties: BaseGitpodAnalyticsEventPropeties = {
-			sessionId: this.sessionId,
-			workspaceId: this.info.getWorkspaceId(),
-			instanceId: this.info.getInstanceId(),
-			appName: vscode.env.appName,
-			uiKind: vscode.env.uiKind === vscode.UIKind.Web ? 'web' : 'desktop',
-			devMode: this.devMode,
-			version: vscode.version,
-			timestamp: Date.now(),
-			'common.extname': this.extension.id,
-			'common.extversion': this.extension.packageJSON.version
-		};
-		const msg: RemoteTrackMessage = {
-			event: eventName,
-			properties: {
-				...baseProperties,
-				...properties,
-			}
-		};
-		if (this.devMode && vscode.env.uiKind === vscode.UIKind.Web) {
-			this.logger.trace(`ANALYTICS: ${JSON.stringify(msg)} `);
-			return Promise.resolve();
-		}
-		try {
-			await this.gitpod.server.trackEvent(msg);
-		} catch (e) {
-			this.logger.error('failed to track event:', e);
-			console.error('failed to track event:', e);
-		}
-	}
-
-	async setPortVisibility(port: number, visibility: PortVisibility): Promise<void> {
-		await this.gitpod.server.openPort(this.info.getWorkspaceId(), {
-			port,
-			visibility
-		});
-	}
-
-	async setTunnelVisibility(port: number, targetPort: number, visibility: TunnelVisiblity): Promise<void> {
-		const request = new TunnelPortRequest();
-		request.setPort(port);
-		request.setTargetPort(targetPort);
-		request.setVisibility(visibility);
-		await util.promisify(this.supervisor.port.tunnel.bind(this.supervisor.port, request, this.supervisor.metadata, {
-			deadline: Date.now() + this.supervisor.deadlines.normal
-		}))();
-	}
-}
+import { GitpodConnection, GitpodExtensionContext, SupervisorConnection } from './gitpodContext';
 
 export async function createGitpodExtensionContext(context: vscode.ExtensionContext): Promise<GitpodExtensionContext | undefined> {
 	const logger = new Log('Gitpod Workspace');
@@ -245,12 +44,12 @@ export async function createGitpodExtensionContext(context: vscode.ExtensionCont
 	const gitpodApi = workspaceInfo.getGitpodApi()!;
 
 	const factory = new JsonRpcProxyFactory<GitpodServer>();
-	const gitpodFunctions: UsedGitpodFunction = ['getWorkspace', 'openPort', 'stopWorkspace', 'setWorkspaceTimeout', 'getWorkspaceTimeout', 'getLoggedInUser', 'takeSnapshot', 'waitForSnapshot', 'controlAdmission', 'sendHeartBeat', 'trackEvent', 'getTeams'];
 	const gitpodService: GitpodConnection = new GitpodServiceImpl<GitpodClient, GitpodServer>(factory.createProxy()) as any;
 	const gitpodScopes = new Set<string>([
 		'resource:workspace::' + workspaceId + '::get/update',
 		'function:accessCodeSyncStorage',
 	]);
+	const gitpodFunctions = ['getWorkspace', 'openPort', 'stopWorkspace', 'setWorkspaceTimeout', 'getWorkspaceTimeout', 'getLoggedInUser', 'takeSnapshot', 'waitForSnapshot', 'controlAdmission', 'sendHeartBeat', 'trackEvent', 'getTeams'];
 	for (const gitpodFunction of gitpodFunctions) {
 		gitpodScopes.add('function:' + gitpodFunction);
 	}
@@ -318,11 +117,9 @@ export async function createGitpodExtensionContext(context: vscode.ExtensionCont
 
 	const ipcHookCli = installCLIProxy(context, logger);
 
-	const config = await import('./gitpod-plugin-model');
 	return new GitpodExtensionContext(
 		context,
 		devMode,
-		config,
 		supervisor,
 		gitpodService,
 		pendignWebSocket,
