+++ extensions/gitpod-web/package.nls.json
@@ -28,5 +28,6 @@
 	"stopSharingWorkspace": "Gitpod: Stop Sharing Running Workspace",
 	"openInStable": "Gitpod: Open in VS Code",
 	"openInInsiders": "Gitpod: Open in VS Code Insiders",
-	"openInBrowser": "Gitpod: Open in Browser"
+	"openInBrowser": "Gitpod: Open in Browser",
+	"buildGitpodYml": "Build container image"
 }
